+++ ref_dx12/q_shwin.cpp
@@ -64,7 +64,7 @@ void *Hunk_Alloc(int size)
 #ifdef VIRTUAL_ALLOC
 	// commit pages as needed
 	//	buf = VirtualAlloc (membase+cursize, size, MEM_COMMIT, PAGE_READWRITE);
-	buf = VirtualAlloc(membase, cursize + size, MEM_COMMIT, PAGE_READWRITE);
+	buf = VirtualAlloc(membase, static_cast<SIZE_T>(cursize) + size, MEM_COMMIT, PAGE_READWRITE);
 	if (!buf)
 	{
 		FormatMessage(FORMAT_MESSAGE_ALLOCATE_BUFFER | FORMAT_MESSAGE_FROM_SYSTEM, NULL, GetLastError(), MAKELANGID(LANG_NEUTRAL, SUBLANG_DEFAULT), (LPTSTR)&buf, 0, NULL);
